Inventory Page Connected
The inventory page now displays all the items in the inventory.
Each item gets a div with the title, description, price, and category, as well as a button that adds the item to the users cart
If a user does not have a cart, one is created and the cart id is stored in the localstorage.